backend/requirements.txt
asgiref==3.7.2
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.3.1
Django==4.2.6
django-cors-headers==4.3.0
django-filter==23.3
django-stubs==4.2.6
django-stubs-ext==4.2.5
djangorestframework==3.14.0
djangorestframework-stubs==3.14.4
drf-flex-fields==1.0.2
drf-spectacular==0.26.5
idna==3.4
inflection==0.5.1
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
Markdown==3.5
mypy==1.6.1
mypy-extensions==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
requests==2.31.0
rpds-py==0.10.6
sqlparse==0.4.4
types-pytz==2023.3.1.1
types-PyYAML==6.0.12.12
types-requests==2.31.0.10
typing_extensions==4.8.0
uritemplate==4.1.1
urllib3==2.0.7
